+++ tests/client/unitTests/admin/PublishApp.js
@@ -10,7 +10,6 @@ describe('PublishApp', function() {
 
   // Load module publish
   beforeEach(function() {
-    module('ngJSONPath');
     module('ov.publish');
   });